Conflicts:
	Makefile.include
	cpu/Makefile

+++ sys/net/net_help/inet_ntop.c
@@ -18,7 +18,7 @@
 #include <string.h>
 #include <stdint.h>
 
-#include "sys/net/destiny/socket.h"
+#include "destiny/socket.h"
 
 #include "inet_ntop.h"
 
@@ -145,7 +145,7 @@ static const char *inet_ntop6(const unsigned char *src, char *dst, size_t size)
      */
     tp = tmp;
 
-    for (i = 0; i < (IN6ADDRSZ / INT16SZ);) {
+    for (i = 0; i < ((int) (IN6ADDRSZ / INT16SZ));) {
         /* Are we inside the best run of 0x00's? */
         if (i == best.base) {
             *tp++ = ':';
